sgpp coding days
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2702 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/combigrid/combischeme/CombiArbitraryScheme.cpp
@@ -9,7 +9,7 @@
 
 combigrid::CombiArbitraryScheme::CombiArbitraryScheme(std::vector<std::vector<
 		int> > level_vectors) :
-	combigrid::CombiSchemeBasis(level_vectors[0].size(), level_vectors[0]) {
+	combigrid::CombiSchemeBasis(static_cast<int>(level_vectors[0].size()), level_vectors[0]) {
 	combigrid::CombigridLevelVector buffer =
 			combigrid::CombigridLevelVector::getCombiLevels(level_vectors);
 	levels_vector_ = buffer.getLevelVec();
